+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## [Unreleased]
+
+### Added
+
+- new pattern expressions: `:bank`, `:cc`, `:pitch-bend`
+
 ## [0.3.0] - 2021-12-24
 
 ### Changed
@@ -16,6 +22,7 @@
 
 Initial release.
 
+[Unreleased]: https://github.com/omkamra/sequencer-midi/compare/0.3.0...HEAD
 [0.3.0]: https://github.com/omkamra/sequencer-midi/compare/0.2.0...0.3.0
 [0.2.0]: https://github.com/omkamra/sequencer-midi/compare/0.1.0...0.2.0
 [0.1.0]: https://github.com/omkamra/sequencer-midi/tree/0.1.0